Upgrade version of loofah (ExamTime/ExamTime#19966)
This is due to the reported security vulnerability.

This was automatically flagged by GitHub, but probably only because we
have published the Gemfile.lock with reference to the insecure version.
The actual version of loofah used by dependent projects will only be
dictated by the gemspec conditions, which do not pin the insecure
version.

In other words, as long as the dependents of `ninja-access` upgrade to a
secure version of loofah they should be fine, regardless of this change
to the `ninja-access` Gemfile.lock.